+++ test/system/promotion/promotion_create_test.rb
@@ -1,0 +1,61 @@
+require 'application_system_test_case'
+
+class PromotionsTest < ApplicationSystemTestCase
+	test 'create promotion' do
+    login_as_user
+    visit root_path
+    click_on 'Promoções'
+    click_on 'Registrar um(a) Promoção'
+    fill_in 'Nome', with: 'Cyber Monday'
+    fill_in 'Descrição', with: 'Promoção de Cyber Monday'
+    fill_in 'Código', with: 'CYBER15'
+    fill_in 'Desconto', with: '15'
+    fill_in 'Quantidade de cupons', with: '90'
+    fill_in 'Data de término', with: '22/12/2033'
+    click_on 'Criar Promoção'
+
+    assert_current_path promotion_path(Promotion.last)
+    assert_text 'Cyber Monday'
+    assert_text 'Promoção de Cyber Monday'
+    assert_text '15,00%'
+    assert_text 'CYBER15'
+    assert_text '22/12/2033'
+    assert_text '90'
+    assert_link 'Voltar'
+  end
+
+  test 'create and attributes cannot be blank' do
+    Promotion.create!(name: 'Natal', description: 'Promoção de Natal',
+                      code: 'NATAL10', discount_rate: 10, coupon_quantity: 100,
+                      expiration_date: '22/12/2033')
+    
+    login_as_user
+    visit root_path
+    click_on 'Promoções'
+    click_on 'Registrar um(a) Promoção'
+    fill_in 'Nome', with: ''
+    fill_in 'Descrição', with: ''
+    fill_in 'Código', with: ''
+    fill_in 'Desconto', with: ''
+    fill_in 'Quantidade de cupons', with: ''
+    fill_in 'Data de término', with: ''
+    click_on 'Criar Promoção'
+
+    assert_text 'não pode ficar em branco', count: 5
+  end
+
+  test 'create and code must be unique' do
+    Promotion.create!(name: 'Natal', description: 'Promoção de Natal',
+                      code: 'NATAL10', discount_rate: 10, coupon_quantity: 100,
+                      expiration_date: '22/12/2033')
+
+    login_as_user
+    visit root_path
+    click_on 'Promoções'
+    click_on 'Registrar um(a) Promoção'
+    fill_in 'Código', with: 'NATAL10'
+    click_on 'Criar Promoção'
+
+    assert_text 'já está em uso'
+  end
+end
