+++ test/system/promotion/promotion_show_test.rb
@@ -1,0 +1,89 @@
+require 'application_system_test_case'
+
+class PromotionsTest < ApplicationSystemTestCase
+	test 'view promotions' do
+    Promotion.create!(name: 'Natal', description: 'Promoção de Natal',
+                      code: 'NATAL10', discount_rate: 10, coupon_quantity: 100,
+                      expiration_date: '22/12/2033')
+    Promotion.create!(name: 'Cyber Monday', coupon_quantity: 100,
+                      description: 'Promoção de Cyber Monday',
+                      code: 'CYBER15', discount_rate: 15,
+                      expiration_date: '22/12/2033')
+    login_as_user
+    visit root_path
+    click_on 'Promoções'
+
+    assert_text 'Natal'
+    assert_text 'Promoção de Natal'
+    assert_text '10,00%'
+    assert_text 'Cyber Monday'
+    assert_text 'Promoção de Cyber Monday'
+    assert_text '15,00%'
+  end
+
+  test 'view promotion details' do
+    Promotion.create!(name: 'Natal', description: 'Promoção de Natal',
+                      code: 'NATAL10', discount_rate: 10, coupon_quantity: 100,
+                      expiration_date: '22/12/2033')
+    Promotion.create!(name: 'Cyber Monday', coupon_quantity: 90,
+                      description: 'Promoção de Cyber Monday',
+                      code: 'CYBER15', discount_rate: 15,
+                      expiration_date: '22/12/2033')
+    login_as_user
+    visit root_path
+    click_on 'Promoções'
+    click_on 'Cyber Monday'
+
+    assert_text 'Cyber Monday'
+    assert_text 'Promoção de Cyber Monday'
+    assert_text '15,00%'
+    assert_text 'CYBER15'
+    assert_text '22/12/2033'
+    assert_text '90'
+  end
+
+  test 'no promotion are available' do
+    login_as_user
+    visit root_path
+    click_on 'Promoções'
+
+    assert_text 'Nenhuma promoção cadastrada'
+  end
+
+  test 'view promotions and return to home page' do
+    Promotion.create!(name: 'Natal', description: 'Promoção de Natal',
+                      code: 'NATAL10', discount_rate: 10, coupon_quantity: 100,
+                      expiration_date: '22/12/2033')
+    login_as_user
+    visit root_path
+    click_on 'Promoções'
+    click_on 'Voltar'
+
+    assert_current_path root_path
+  end
+
+  test 'view details and return to promotions page' do
+    Promotion.create!(name: 'Natal', description: 'Promoção de Natal',
+                      code: 'NATAL10', discount_rate: 10, coupon_quantity: 100,
+                      expiration_date: '22/12/2033')
+    login_as_user
+    visit root_path
+    click_on 'Promoções'
+    click_on 'Natal'
+    click_on 'Voltar'
+
+    assert_current_path promotions_path
+  end
+
+  test 'do not view promotion link without login' do
+    visit root_path
+
+    assert_no_link 'Promoções'
+  end
+
+  test 'verify access to show is disabled' do
+    visit promotions_path
+
+    assert_current_path new_user_session_path
+  end
+end
