+++ test/system/promotion/promotion_update_test.rb
@@ -1,0 +1,90 @@
+require 'application_system_test_case'
+
+class PromotionsTest < ApplicationSystemTestCase
+	test 'update promotions' do
+    promotion = Promotion.create!(name: 'Natal',
+                                  description: 'Promoção de Natal',
+                                  code: 'NATAL10', discount_rate: 10,
+                                  coupon_quantity: 100,
+                                  expiration_date: '22/12/2033')
+
+    login_as_user
+    visit root_path
+    click_on 'Promoções'
+    click_on promotion.name
+    click_on 'Editar'
+    fill_in 'Nome', with: 'Cyber Monday'
+    fill_in 'Descrição', with: 'Promoção de Cyber Monday'
+    fill_in 'Código', with: 'CYBER15'
+    fill_in 'Desconto', with: '15'
+    fill_in 'Quantidade de cupons', with: '90'
+    fill_in 'Data de término', with: '22/12/2033'
+    click_on 'Atualizar Promoção'
+
+    assert_text 'Cyber Monday'
+    assert_text 'Promoção de Cyber Monday'
+    assert_text '15,00%'
+    assert_text 'CYBER15'
+    assert_text '22/12/2033'
+    assert_text '90'
+    assert_link 'Voltar'
+  end
+  
+  test 'update and attributes cannot be blank' do
+    promotion = Promotion.create!(name: 'Natal',
+                                  description: 'Promoção de Natal',
+                                  code: 'NATAL10', discount_rate: 10,
+                                  coupon_quantity: 100,
+                                  expiration_date: '22/12/2033')
+
+    login_as_user
+    visit root_path
+    click_on 'Promoções'
+    click_on promotion.name
+    click_on 'Editar'
+    fill_in 'Nome', with: ''
+    fill_in 'Descrição', with: ''
+    fill_in 'Código', with: ''
+    fill_in 'Desconto', with: ''
+    fill_in 'Quantidade de cupons', with: ''
+    fill_in 'Data de término', with: ''
+    click_on 'Atualizar Promoção'
+
+    assert_text 'não pode ficar em branco', count: 5
+  end
+
+  test 'update and code/name must be unique' do
+    promotion = Promotion.create!(name: 'Natal',
+                                  description: 'Promoção de Natal',
+                                  code: 'NATAL10', discount_rate: 10,
+                                  coupon_quantity: 100,
+                                  expiration_date: '22/12/2033')
+    Promotion.create!(name: 'Cyber Monday', coupon_quantity: 100,
+                      description: 'Promoção de Cyber Monday',
+                      code: 'CYBER15', discount_rate: 15,
+                      expiration_date: '22/12/2033')
+
+    login_as_user
+    visit root_path
+    click_on 'Promoções'
+    click_on promotion.name
+    click_on 'Editar'
+    fill_in 'Nome', with: 'Cyber Monday'
+    fill_in 'Código', with: 'CYBER15'
+    click_on 'Atualizar Promoção'
+
+    assert_text 'já está em uso', count: 2
+  end
+
+  test 'can not edit promotion without login' do
+    promotion = Promotion.create!(name: 'Natal',
+                                  description: 'Promoção de Natal',
+                                  code: 'NATAL10', discount_rate: 10,
+                                  coupon_quantity: 100,
+                                  expiration_date: '22/12/2033')
+
+    visit edit_promotion_path(promotion)
+
+    assert_current_path new_user_session_path
+  end
+end
